requirements.txt
astroid==2.3.3
autopep8==1.5.3
click==7.1.2
colorama==0.4.1
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.10
imageio==2.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
networkx==2.5
numpy==1.20.1
Pillow==8.1.1
pycodestyle==2.6.0
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.13.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2021.3.4
toml==0.10.1
typed-ast==1.4.0
Werkzeug==1.0.1
wrapt==1.11.2
